fix update data in left menu after update

+++ app/src/main/java/com/android/template/data/models/api/model/UserModel.kt
@@ -7,7 +7,7 @@ data class UserModel(
     val email: String,
 
     @Expose
-    val profileId: Int,
+    val profileId: Long,
 
     @Expose
     val avatarUrl: String,
